Explicit type conversion

+++ src/dataset.py
@@ -166,12 +166,12 @@ class BaseDataset(CocoDetection):
         target['labels'] = np.array(target['labels'], dtype=np.int64)
 
         if 'boxes' in target:
-            target['boxes'] = np.array(target['boxes'], dtype=np.float)
+            target['boxes'] = np.array(target['boxes'], dtype=np.float32)
 
         if 'masks' in target:
             target['masks'] = np.array(target['masks'], dtype=np.uint8)
 
-        target['area'] = np.array(target['area'], dtype=np.float)
+        target['area'] = np.array(target['area'], dtype=np.float32)
 
     def category_name(self, category_id):
         """Return category name given its id."""
